Introducing Menu and Overlay
The Menu is just laid plainly over anything underlying, and expects something to be there when start is pressed, because that just makes it invisible.

This means for getting back to the menu, we only need to reset the playerposition (as in the example project) and make it visible again.

"Start" is emitted on the start button being clicked
"Settings" is emitted when the "Settings" button is clicked
"Exit" is emitted when the exit button is clicked.

Buttons can be easily added, they just need a name